ELO algorithm was implemented totally wrong. Added tests and fixed the implementation.
Also changes thr ranking to be a float so that rounding errors don't grow too large.

+++ lib/models.py
@@ -15,7 +15,7 @@ class Player(db.Model):
   user = db.UserProperty(required=True)
   pseudonym = db.StringProperty()
   location = db.GeoPtProperty(default=DC['geoPt'])
-  rank = db.IntegerProperty(default=INITIAL_RANK)
+  rank = db.FloatProperty(default=INITIAL_RANK)
 
   def __str__(self):
     return "[Psuedonym: %s, Nickname: %s]" % (self.pseudonym, self.user.nickname())
